requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
MouseInfo==0.1.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python3-xlib==0.15
PyTweening==1.0.3
six==1.15.0
